Enable testcontainers support, remove second version of pg-embedded
