+++ .github/workflows/gradle.yml
@@ -22,4 +22,4 @@ jobs:
                 with:
                     java-version: 1.8
             -   name: Build with Gradle
-                run: ./gradlew build
+                run: ./gradlew build -PpostgresType=testContainersPostgres
