+++ build.gradle
@@ -16,6 +16,18 @@ repositories {
     mavenCentral()
 }
 
+def postgresType = project.properties.getOrDefault('postgresType', 'opentablePostgres')
+def pickPostgres = { embeddedPgOption, containerPgOption ->
+    switch (postgresType) {
+        case 'opentablePostgres':
+            return embeddedPgOption
+        case 'testContainersPostgres':
+            return containerPgOption
+        default:
+            throw new GradleException("Unsupported postgres type: $postgresType")
+    }
+}
+
 dependencies {
     compile 'com.google.code.findbugs:jsr305:3.0.2'
     compile 'org.apache.commons:commons-lang3:3.9'
@@ -26,17 +38,45 @@ dependencies {
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.5.2'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.5.2'
     testImplementation 'org.junit.jupiter:junit-jupiter-params:5.5.2'
-    testImplementation 'com.opentable.components:otj-pg-embedded:0.13.3'
     testImplementation 'org.hamcrest:hamcrest:2.2'
     testImplementation 'ch.qos.logback:logback-classic:1.2.3'
     testImplementation 'org.mockito:mockito-core:3.2.0'
-    testImplementation 'io.zonky.test:embedded-postgres:1.2.6'
-    testImplementation enforcedPlatform('io.zonky.test.postgres:embedded-postgres-binaries-bom:11.5.0')
+
+    pickPostgres(
+            [
+                    'com.opentable.components:otj-pg-embedded:0.13.3'
+            ],
+            [
+                    'org.testcontainers:testcontainers:1.12.5',
+                    'org.testcontainers:junit-jupiter:1.12.5',
+                    'org.testcontainers:postgresql:1.12.5',
+                    'org.postgresql:postgresql:42.2.10'
+            ]
+    ).each { testImplementation it }
+}
+
+sourceSets {
+    test {
+        java {
+            srcDir pickPostgres('src/embedded-pg/java', 'src/testcontainers-pg/java')
+        }
+    }
 }
 
 test {
     useJUnitPlatform()
 }
+
+//def integrationTest = tasks.register('integrationTest', Test) {
+//    description = 'Runs integration tests.'
+//    group = 'verification'
+//
+//    testClassesDirs = sourceSets.integrationTest.output.classesDirs
+//    classpath = sourceSets.integrationTest.runtimeClasspath
+//    shouldRunAfter('test')
+//}
+
+//tasks.named('check') { dependsOn(integrationTest) }
 
 jar {
     manifest {
