+++ gradle.properties
@@ -1,3 +1,4 @@
 sonatypeUsername=
 sonatypePassword=
 systemProp.org.gradle.internal.publish.checksums.insecure=true
+postgresType=testContainersPostgres
