+++ src/test/java/io/github/mfvanek/pg/index/health/IndexesHealthImplTest.java
@@ -5,17 +5,20 @@
  * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
  */
 
-package io.github.mfvanek.pg.index.maintenance;
+package io.github.mfvanek.pg.index.health;
 
-import io.github.mfvanek.pg.connection.PgConnection;
+import io.github.mfvanek.pg.EmbeddedPostgresExtension;
+import io.github.mfvanek.pg.PreparedDbExtension;
+import io.github.mfvanek.pg.connection.HighAvailabilityPgConnection;
+import io.github.mfvanek.pg.connection.HighAvailabilityPgConnectionImpl;
 import io.github.mfvanek.pg.connection.PgConnectionImpl;
+import io.github.mfvanek.pg.index.maintenance.MaintenanceFactoryImpl;
 import io.github.mfvanek.pg.model.DuplicatedIndexes;
 import io.github.mfvanek.pg.model.ForeignKey;
 import io.github.mfvanek.pg.model.Index;
 import io.github.mfvanek.pg.model.IndexWithBloat;
 import io.github.mfvanek.pg.model.IndexWithNulls;
 import io.github.mfvanek.pg.model.IndexWithSize;
-import io.github.mfvanek.pg.model.PgContext;
 import io.github.mfvanek.pg.model.Table;
 import io.github.mfvanek.pg.model.TableWithBloat;
 import io.github.mfvanek.pg.model.TableWithMissingIndex;
@@ -23,6 +26,7 @@ import io.github.mfvanek.pg.model.UnusedIndex;
 import io.github.mfvanek.pg.utils.DatabaseAwareTestBase;
 import io.github.mfvanek.pg.utils.DatabasePopulator;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -42,19 +46,23 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
+public final class IndexesHealthImplTest extends DatabaseAwareTestBase {
+    @RegisterExtension
+    static final PreparedDbExtension embeddedPostgres =
+            EmbeddedPostgresExtension.preparedDatabase();
 
-    private final IndexMaintenance indexMaintenance;
+    private final IndexesHealth indexesHealth;
 
-    IndexMaintenanceImplTestBase(@Nonnull final DataSource dataSource) {
-        super(dataSource);
-        final PgConnection pgConnection = PgConnectionImpl.ofMaster(dataSource);
-        this.indexMaintenance = new IndexMaintenanceImpl(pgConnection);
+    IndexesHealthImplTest() {
+        super(embeddedPostgres.getTestDatabase());
+        final HighAvailabilityPgConnection haPgConnection = HighAvailabilityPgConnectionImpl.of(
+                PgConnectionImpl.ofMaster(embeddedPostgres.getTestDatabase()));
+        this.indexesHealth = new IndexesHealthImpl(haPgConnection, new MaintenanceFactoryImpl());
     }
 
     @Test
-    void getInvalidIndexesOnEmptyDataBase() {
-        final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes();
+    void getInvalidIndexesOnEmptyDatabase() {
+        final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes();
         assertNotNull(invalidIndexes);
         assertEquals(0, invalidIndexes.size());
     }
@@ -65,7 +73,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes(ctx);
+                    final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes(ctx);
                     assertNotNull(invalidIndexes);
                     assertEquals(0, invalidIndexes.size());
                 });
@@ -77,7 +85,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withInvalidIndex(),
                 ctx -> {
-                    final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes(ctx);
+                    final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes(ctx);
                     assertNotNull(invalidIndexes);
                     assertEquals(1, invalidIndexes.size());
                     final Index index = invalidIndexes.get(0);
@@ -92,8 +100,8 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getDuplicatedIndexesOnEmptyDataBase() {
-        final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes();
+    void getDuplicatedIndexesOnEmptyDatabase() {
+        final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes();
         assertNotNull(duplicatedIndexes);
         assertEquals(0, duplicatedIndexes.size());
     }
@@ -104,7 +112,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes(ctx);
+                    final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes(ctx);
                     assertNotNull(duplicatedIndexes);
                     assertEquals(0, duplicatedIndexes.size());
                 });
@@ -114,9 +122,9 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     @ValueSource(strings = {"public", "custom"})
     void getDuplicatedIndexesOnDatabaseWithThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
-                dbp -> dbp.withReferences().withDuplicatedIndex(),
+                dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
                 ctx -> {
-                    final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes(ctx);
+                    final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes(ctx);
                     assertNotNull(duplicatedIndexes);
                     assertEquals(1, duplicatedIndexes.size());
                     final DuplicatedIndexes entry = duplicatedIndexes.get(0);
@@ -144,8 +152,8 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getIntersectedIndexesOnEmptyDataBase() {
-        final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes();
+    void getIntersectedIndexesOnEmptyDatabase() {
+        final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes();
         assertNotNull(intersectedIndexes);
         assertEquals(0, intersectedIndexes.size());
     }
@@ -156,7 +164,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes(ctx);
+                    final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes(ctx);
                     assertNotNull(intersectedIndexes);
                     assertEquals(0, intersectedIndexes.size());
                 });
@@ -168,7 +176,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
                 ctx -> {
-                    final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes(ctx);
+                    final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes(ctx);
                     assertNotNull(intersectedIndexes);
                     assertEquals(1, intersectedIndexes.size());
                     final DuplicatedIndexes entry = intersectedIndexes.get(0);
@@ -196,19 +204,19 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getPotentiallyUnusedIndexesOnEmptyDataBase() {
-        final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes();
+    void getUnusedIndexesOnEmptyDatabase() {
+        final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes();
         assertNotNull(unusedIndexes);
         assertEquals(0, unusedIndexes.size());
     }
 
     @ParameterizedTest
     @ValueSource(strings = {"public", "custom"})
-    void getPotentiallyUnusedIndexesOnDatabaseWithoutThem(final String schemaName) {
+    void getUnusedIndexesOnDatabaseWithoutThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes(ctx);
+                    final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes(ctx);
                     assertNotNull(unusedIndexes);
                     assertEquals(0, unusedIndexes.size());
                 });
@@ -216,11 +224,11 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
 
     @ParameterizedTest
     @ValueSource(strings = {"public", "custom"})
-    void getPotentiallyUnusedIndexesOnDatabaseWithThem(final String schemaName) {
+    void getUnusedIndexesOnDatabaseWithThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
                 ctx -> {
-                    final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes(ctx);
+                    final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes(ctx);
                     assertNotNull(unusedIndexes);
                     assertThat(unusedIndexes.size(), equalTo(3));
                     final Set<String> names = unusedIndexes.stream().map(UnusedIndex::getIndexName).collect(toSet());
@@ -239,8 +247,8 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getForeignKeysNotCoveredWithIndexOnEmptyDataBase() {
-        final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex();
+    void getForeignKeysNotCoveredWithIndexOnEmptyDatabase() {
+        final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex();
         assertNotNull(foreignKeys);
         assertEquals(0, foreignKeys.size());
     }
@@ -251,7 +259,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp,
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(0, foreignKeys.size());
                 });
@@ -263,7 +271,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(1, foreignKeys.size());
                     final ForeignKey foreignKey = foreignKeys.get(0);
@@ -282,7 +290,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withNonSuitableIndex(),
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(1, foreignKeys.size());
                     final ForeignKey foreignKey = foreignKeys.get(0);
@@ -301,17 +309,17 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withSuitableIndex(),
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(0, foreignKeys.size());
                 });
     }
 
     @Test
-    void getTablesWithMissingIndexesOnEmptyDataBase() {
-        final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes();
+    void getTablesWithMissingIndexesOnEmptyDatabase() {
+        final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes();
         assertNotNull(tables);
-        assertEquals(0, tables.size());
+        assertThat(tables, hasSize(0));
     }
 
     @ParameterizedTest
@@ -320,9 +328,9 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes(ctx);
+                    final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes(ctx);
                     assertNotNull(tables);
-                    assertEquals(0, tables.size());
+                    assertThat(tables, hasSize(0));
                 });
     }
 
@@ -333,7 +341,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
                     tryToFindAccountByClientId(schemaName);
-                    final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes(ctx);
+                    final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes(ctx);
                     assertNotNull(tables);
                     assertThat(tables, hasSize(1));
                     final TableWithMissingIndex table = tables.get(0);
@@ -348,8 +356,8 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getTablesWithoutPrimaryKeyOnEmptyDataBase() {
-        final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey();
+    void getTablesWithoutPrimaryKeyOnEmptyDatabase() {
+        final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey();
         assertNotNull(tables);
         assertEquals(0, tables.size());
     }
@@ -360,7 +368,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey(ctx);
+                    final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey(ctx);
                     assertNotNull(tables);
                     assertEquals(0, tables.size());
                 });
@@ -372,7 +380,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withTableWithoutPrimaryKey(),
                 ctx -> {
-                    final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey(ctx);
+                    final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey(ctx);
                     assertNotNull(tables);
                     assertThat(tables, hasSize(1));
                     final Table table = tables.get(0);
@@ -385,8 +393,8 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getIndexesWithNullValuesOnEmptyDataBase() {
-        final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues();
+    void getIndexesWithNullValuesOnEmptyDatabase() {
+        final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues();
         assertNotNull(indexes);
         assertEquals(0, indexes.size());
     }
@@ -397,7 +405,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues(ctx);
+                    final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues(ctx);
                     assertNotNull(indexes);
                     assertEquals(0, indexes.size());
                 });
@@ -409,7 +417,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withNullValuesInIndex(),
                 ctx -> {
-                    final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues(ctx);
+                    final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues(ctx);
                     assertNotNull(indexes);
                     assertEquals(1, indexes.size());
                     final IndexWithNulls indexWithNulls = indexes.get(0);
@@ -424,7 +432,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
 
     @Test
     void getIndexesWithBloatOnEmptyDataBase() {
-        final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat();
+        final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat();
         assertNotNull(indexes);
         assertEquals(0, indexes.size());
     }
@@ -436,7 +444,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withStatistics(),
                 ctx -> {
                     waitForStatisticsCollector();
-                    final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat(ctx);
+                    final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat(ctx);
                     assertNotNull(indexes);
                     assertEquals(0, indexes.size());
                 });
@@ -451,7 +459,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                     waitForStatisticsCollector();
                     assertTrue(existsStatisticsForTable(ctx, "accounts"));
 
-                    final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat(ctx);
+                    final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat(ctx);
                     assertNotNull(indexes);
                     assertEquals(3, indexes.size());
                     final IndexWithBloat index = indexes.get(0);
@@ -470,7 +478,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
 
     @Test
     void getTablesWithBloatOnEmptyDataBase() {
-        final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat();
+        final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat();
         assertNotNull(tables);
         assertEquals(0, tables.size());
     }
@@ -482,7 +490,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withStatistics(),
                 ctx -> {
                     waitForStatisticsCollector();
-                    final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat(ctx);
+                    final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat(ctx);
                     assertNotNull(tables);
                     assertEquals(0, tables.size());
                 });
@@ -497,7 +505,7 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                     waitForStatisticsCollector();
                     assertTrue(existsStatisticsForTable(ctx, "accounts"));
 
-                    final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat(ctx);
+                    final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat(ctx);
                     assertNotNull(tables);
                     assertEquals(2, tables.size());
                     final TableWithBloat table = tables.get(0);
@@ -512,26 +520,17 @@ abstract class IndexMaintenanceImplTestBase extends DatabaseAwareTestBase {
                 });
     }
 
-    @Test
-    void securityTest() {
-        executeTestOnDatabase("public",
-                dbp -> dbp.withReferences().withData().withNullValuesInIndex(),
+    @ParameterizedTest
+    @ValueSource(strings = {"public", "custom"})
+    void shouldResetCounters(final String schemaName) {
+        executeTestOnDatabase(schemaName,
+                dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final long before = getRowsCount(ctx.getSchemaName(), "clients");
-                    assertEquals(1001L, before);
-                    List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues
-                            (PgContext.of("; truncate table clients;"));
-                    assertNotNull(indexes);
-                    assertThat(indexes, hasSize(0));
-                    assertEquals(before, getRowsCount(ctx.getSchemaName(), "clients"));
-
-                    indexes = indexMaintenance.getIndexesWithNullValues(PgContext.of("; select pg_sleep(100000000);"));
-                    assertNotNull(indexes);
-                    assertThat(indexes, hasSize(0));
-
-                    indexes = indexMaintenance.getIndexesWithNullValues(ctx);
-                    assertNotNull(indexes);
-                    assertThat(indexes, hasSize(1));
+                    tryToFindAccountByClientId(schemaName);
+                    assertThat(getSeqScansForAccounts(ctx), greaterThanOrEqualTo(AMOUNT_OF_TRIES));
+                    indexesHealth.resetStatistics();
+                    waitForStatisticsCollector();
+                    assertEquals(0L, getSeqScansForAccounts(ctx));
                 });
     }
 }
