--- src/test/java/io/github/mfvanek/pg/index/health/IndexesHealthImplTestPg11.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.index.health;
-
-import io.zonky.test.db.postgres.junit5.EmbeddedPostgresExtension;
-import io.zonky.test.db.postgres.junit5.PreparedDbExtension;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
-
-class IndexesHealthImplTestPg11 extends IndexesHealthImplTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension.preparedDatabase(ds -> {});
-
-    IndexesHealthImplTestPg11() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-
-    @Test
-    void pgVersion() {
-        assertThat(getPgVersion(), containsString("PostgreSQL 11.5"));
-    }
-}
