--- src/test/java/io/github/mfvanek/pg/index/health/logger/IndexesHealthLoggerTestPg10.java
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.index.health.logger;
-
-import com.opentable.db.postgres.junit5.EmbeddedPostgresExtension;
-import com.opentable.db.postgres.junit5.PreparedDbExtension;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-class IndexesHealthLoggerTestPg10 extends IndexesHealthLoggerTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension.preparedDatabase(ds -> {});
-
-    IndexesHealthLoggerTestPg10() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-}
