+++ src/test/java/io/github/mfvanek/pg/index/maintenance/IndexMaintenanceImplTest.java
@@ -5,18 +5,19 @@
  * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
  */
 
-package io.github.mfvanek.pg.index.health;
+package io.github.mfvanek.pg.index.maintenance;
 
-import io.github.mfvanek.pg.connection.HighAvailabilityPgConnection;
-import io.github.mfvanek.pg.connection.HighAvailabilityPgConnectionImpl;
+import io.github.mfvanek.pg.EmbeddedPostgresExtension;
+import io.github.mfvanek.pg.PreparedDbExtension;
+import io.github.mfvanek.pg.connection.PgConnection;
 import io.github.mfvanek.pg.connection.PgConnectionImpl;
-import io.github.mfvanek.pg.index.maintenance.MaintenanceFactoryImpl;
 import io.github.mfvanek.pg.model.DuplicatedIndexes;
 import io.github.mfvanek.pg.model.ForeignKey;
 import io.github.mfvanek.pg.model.Index;
 import io.github.mfvanek.pg.model.IndexWithBloat;
 import io.github.mfvanek.pg.model.IndexWithNulls;
 import io.github.mfvanek.pg.model.IndexWithSize;
+import io.github.mfvanek.pg.model.PgContext;
 import io.github.mfvanek.pg.model.Table;
 import io.github.mfvanek.pg.model.TableWithBloat;
 import io.github.mfvanek.pg.model.TableWithMissingIndex;
@@ -24,11 +25,10 @@ import io.github.mfvanek.pg.model.UnusedIndex;
 import io.github.mfvanek.pg.utils.DatabaseAwareTestBase;
 import io.github.mfvanek.pg.utils.DatabasePopulator;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import javax.annotation.Nonnull;
-import javax.sql.DataSource;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -43,20 +43,22 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
+public final class IndexMaintenanceImplTest extends DatabaseAwareTestBase {
+    @RegisterExtension
+    static final PreparedDbExtension embeddedPostgres =
+            EmbeddedPostgresExtension.preparedDatabase();
 
-    private final IndexesHealth indexesHealth;
+    private final IndexMaintenance indexMaintenance;
 
-    IndexesHealthImplTestBase(@Nonnull final DataSource dataSource) {
-        super(dataSource);
-        final HighAvailabilityPgConnection haPgConnection = HighAvailabilityPgConnectionImpl.of(
-                PgConnectionImpl.ofMaster(dataSource));
-        this.indexesHealth = new IndexesHealthImpl(haPgConnection, new MaintenanceFactoryImpl());
+    IndexMaintenanceImplTest() {
+        super(embeddedPostgres.getTestDatabase());
+        final PgConnection pgConnection = PgConnectionImpl.ofMaster(embeddedPostgres.getTestDatabase());
+        this.indexMaintenance = new IndexMaintenanceImpl(pgConnection);
     }
 
     @Test
-    void getInvalidIndexesOnEmptyDatabase() {
-        final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes();
+    void getInvalidIndexesOnEmptyDataBase() {
+        final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes();
         assertNotNull(invalidIndexes);
         assertEquals(0, invalidIndexes.size());
     }
@@ -67,7 +69,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes(ctx);
+                    final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes(ctx);
                     assertNotNull(invalidIndexes);
                     assertEquals(0, invalidIndexes.size());
                 });
@@ -79,7 +81,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withInvalidIndex(),
                 ctx -> {
-                    final List<Index> invalidIndexes = indexesHealth.getInvalidIndexes(ctx);
+                    final List<Index> invalidIndexes = indexMaintenance.getInvalidIndexes(ctx);
                     assertNotNull(invalidIndexes);
                     assertEquals(1, invalidIndexes.size());
                     final Index index = invalidIndexes.get(0);
@@ -94,8 +96,8 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getDuplicatedIndexesOnEmptyDatabase() {
-        final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes();
+    void getDuplicatedIndexesOnEmptyDataBase() {
+        final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes();
         assertNotNull(duplicatedIndexes);
         assertEquals(0, duplicatedIndexes.size());
     }
@@ -106,7 +108,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes(ctx);
+                    final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes(ctx);
                     assertNotNull(duplicatedIndexes);
                     assertEquals(0, duplicatedIndexes.size());
                 });
@@ -116,9 +118,9 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     @ValueSource(strings = {"public", "custom"})
     void getDuplicatedIndexesOnDatabaseWithThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
-                dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
+                dbp -> dbp.withReferences().withDuplicatedIndex(),
                 ctx -> {
-                    final List<DuplicatedIndexes> duplicatedIndexes = indexesHealth.getDuplicatedIndexes(ctx);
+                    final List<DuplicatedIndexes> duplicatedIndexes = indexMaintenance.getDuplicatedIndexes(ctx);
                     assertNotNull(duplicatedIndexes);
                     assertEquals(1, duplicatedIndexes.size());
                     final DuplicatedIndexes entry = duplicatedIndexes.get(0);
@@ -146,8 +148,8 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getIntersectedIndexesOnEmptyDatabase() {
-        final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes();
+    void getIntersectedIndexesOnEmptyDataBase() {
+        final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes();
         assertNotNull(intersectedIndexes);
         assertEquals(0, intersectedIndexes.size());
     }
@@ -158,7 +160,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes(ctx);
+                    final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes(ctx);
                     assertNotNull(intersectedIndexes);
                     assertEquals(0, intersectedIndexes.size());
                 });
@@ -170,7 +172,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
                 ctx -> {
-                    final List<DuplicatedIndexes> intersectedIndexes = indexesHealth.getIntersectedIndexes(ctx);
+                    final List<DuplicatedIndexes> intersectedIndexes = indexMaintenance.getIntersectedIndexes(ctx);
                     assertNotNull(intersectedIndexes);
                     assertEquals(1, intersectedIndexes.size());
                     final DuplicatedIndexes entry = intersectedIndexes.get(0);
@@ -198,19 +200,19 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getUnusedIndexesOnEmptyDatabase() {
-        final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes();
+    void getPotentiallyUnusedIndexesOnEmptyDataBase() {
+        final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes();
         assertNotNull(unusedIndexes);
         assertEquals(0, unusedIndexes.size());
     }
 
     @ParameterizedTest
     @ValueSource(strings = {"public", "custom"})
-    void getUnusedIndexesOnDatabaseWithoutThem(final String schemaName) {
+    void getPotentiallyUnusedIndexesOnDatabaseWithoutThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes(ctx);
+                    final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes(ctx);
                     assertNotNull(unusedIndexes);
                     assertEquals(0, unusedIndexes.size());
                 });
@@ -218,11 +220,11 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
 
     @ParameterizedTest
     @ValueSource(strings = {"public", "custom"})
-    void getUnusedIndexesOnDatabaseWithThem(final String schemaName) {
+    void getPotentiallyUnusedIndexesOnDatabaseWithThem(final String schemaName) {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withDuplicatedIndex(),
                 ctx -> {
-                    final List<UnusedIndex> unusedIndexes = indexesHealth.getUnusedIndexes(ctx);
+                    final List<UnusedIndex> unusedIndexes = indexMaintenance.getPotentiallyUnusedIndexes(ctx);
                     assertNotNull(unusedIndexes);
                     assertThat(unusedIndexes.size(), equalTo(3));
                     final Set<String> names = unusedIndexes.stream().map(UnusedIndex::getIndexName).collect(toSet());
@@ -241,8 +243,8 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getForeignKeysNotCoveredWithIndexOnEmptyDatabase() {
-        final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex();
+    void getForeignKeysNotCoveredWithIndexOnEmptyDataBase() {
+        final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex();
         assertNotNull(foreignKeys);
         assertEquals(0, foreignKeys.size());
     }
@@ -253,7 +255,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp,
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(0, foreignKeys.size());
                 });
@@ -265,7 +267,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 DatabasePopulator::withReferences,
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(1, foreignKeys.size());
                     final ForeignKey foreignKey = foreignKeys.get(0);
@@ -284,7 +286,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withNonSuitableIndex(),
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(1, foreignKeys.size());
                     final ForeignKey foreignKey = foreignKeys.get(0);
@@ -303,17 +305,17 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withSuitableIndex(),
                 ctx -> {
-                    final List<ForeignKey> foreignKeys = indexesHealth.getForeignKeysNotCoveredWithIndex(ctx);
+                    final List<ForeignKey> foreignKeys = indexMaintenance.getForeignKeysNotCoveredWithIndex(ctx);
                     assertNotNull(foreignKeys);
                     assertEquals(0, foreignKeys.size());
                 });
     }
 
     @Test
-    void getTablesWithMissingIndexesOnEmptyDatabase() {
-        final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes();
+    void getTablesWithMissingIndexesOnEmptyDataBase() {
+        final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes();
         assertNotNull(tables);
-        assertThat(tables, hasSize(0));
+        assertEquals(0, tables.size());
     }
 
     @ParameterizedTest
@@ -322,9 +324,9 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes(ctx);
+                    final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes(ctx);
                     assertNotNull(tables);
-                    assertThat(tables, hasSize(0));
+                    assertEquals(0, tables.size());
                 });
     }
 
@@ -335,7 +337,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
                     tryToFindAccountByClientId(schemaName);
-                    final List<TableWithMissingIndex> tables = indexesHealth.getTablesWithMissingIndexes(ctx);
+                    final List<TableWithMissingIndex> tables = indexMaintenance.getTablesWithMissingIndexes(ctx);
                     assertNotNull(tables);
                     assertThat(tables, hasSize(1));
                     final TableWithMissingIndex table = tables.get(0);
@@ -350,8 +352,8 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getTablesWithoutPrimaryKeyOnEmptyDatabase() {
-        final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey();
+    void getTablesWithoutPrimaryKeyOnEmptyDataBase() {
+        final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey();
         assertNotNull(tables);
         assertEquals(0, tables.size());
     }
@@ -362,7 +364,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey(ctx);
+                    final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey(ctx);
                     assertNotNull(tables);
                     assertEquals(0, tables.size());
                 });
@@ -374,7 +376,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withTableWithoutPrimaryKey(),
                 ctx -> {
-                    final List<Table> tables = indexesHealth.getTablesWithoutPrimaryKey(ctx);
+                    final List<Table> tables = indexMaintenance.getTablesWithoutPrimaryKey(ctx);
                     assertNotNull(tables);
                     assertThat(tables, hasSize(1));
                     final Table table = tables.get(0);
@@ -387,8 +389,8 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
     }
 
     @Test
-    void getIndexesWithNullValuesOnEmptyDatabase() {
-        final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues();
+    void getIndexesWithNullValuesOnEmptyDataBase() {
+        final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues();
         assertNotNull(indexes);
         assertEquals(0, indexes.size());
     }
@@ -399,7 +401,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData(),
                 ctx -> {
-                    final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues(ctx);
+                    final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues(ctx);
                     assertNotNull(indexes);
                     assertEquals(0, indexes.size());
                 });
@@ -411,7 +413,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
         executeTestOnDatabase(schemaName,
                 dbp -> dbp.withReferences().withData().withNullValuesInIndex(),
                 ctx -> {
-                    final List<IndexWithNulls> indexes = indexesHealth.getIndexesWithNullValues(ctx);
+                    final List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues(ctx);
                     assertNotNull(indexes);
                     assertEquals(1, indexes.size());
                     final IndexWithNulls indexWithNulls = indexes.get(0);
@@ -426,7 +428,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
 
     @Test
     void getIndexesWithBloatOnEmptyDataBase() {
-        final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat();
+        final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat();
         assertNotNull(indexes);
         assertEquals(0, indexes.size());
     }
@@ -438,7 +440,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withStatistics(),
                 ctx -> {
                     waitForStatisticsCollector();
-                    final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat(ctx);
+                    final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat(ctx);
                     assertNotNull(indexes);
                     assertEquals(0, indexes.size());
                 });
@@ -453,7 +455,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                     waitForStatisticsCollector();
                     assertTrue(existsStatisticsForTable(ctx, "accounts"));
 
-                    final List<IndexWithBloat> indexes = indexesHealth.getIndexesWithBloat(ctx);
+                    final List<IndexWithBloat> indexes = indexMaintenance.getIndexesWithBloat(ctx);
                     assertNotNull(indexes);
                     assertEquals(3, indexes.size());
                     final IndexWithBloat index = indexes.get(0);
@@ -472,7 +474,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
 
     @Test
     void getTablesWithBloatOnEmptyDataBase() {
-        final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat();
+        final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat();
         assertNotNull(tables);
         assertEquals(0, tables.size());
     }
@@ -484,7 +486,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                 dbp -> dbp.withReferences().withStatistics(),
                 ctx -> {
                     waitForStatisticsCollector();
-                    final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat(ctx);
+                    final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat(ctx);
                     assertNotNull(tables);
                     assertEquals(0, tables.size());
                 });
@@ -499,7 +501,7 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                     waitForStatisticsCollector();
                     assertTrue(existsStatisticsForTable(ctx, "accounts"));
 
-                    final List<TableWithBloat> tables = indexesHealth.getTablesWithBloat(ctx);
+                    final List<TableWithBloat> tables = indexMaintenance.getTablesWithBloat(ctx);
                     assertNotNull(tables);
                     assertEquals(2, tables.size());
                     final TableWithBloat table = tables.get(0);
@@ -514,17 +516,26 @@ abstract class IndexesHealthImplTestBase extends DatabaseAwareTestBase {
                 });
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {"public", "custom"})
-    void shouldResetCounters(final String schemaName) {
-        executeTestOnDatabase(schemaName,
-                dbp -> dbp.withReferences().withData(),
+    @Test
+    void securityTest() {
+        executeTestOnDatabase("public",
+                dbp -> dbp.withReferences().withData().withNullValuesInIndex(),
                 ctx -> {
-                    tryToFindAccountByClientId(schemaName);
-                    assertThat(getSeqScansForAccounts(ctx), greaterThanOrEqualTo(AMOUNT_OF_TRIES));
-                    indexesHealth.resetStatistics();
-                    waitForStatisticsCollector();
-                    assertEquals(0L, getSeqScansForAccounts(ctx));
+                    final long before = getRowsCount(ctx.getSchemaName(), "clients");
+                    assertEquals(1001L, before);
+                    List<IndexWithNulls> indexes = indexMaintenance.getIndexesWithNullValues
+                            (PgContext.of("; truncate table clients;"));
+                    assertNotNull(indexes);
+                    assertThat(indexes, hasSize(0));
+                    assertEquals(before, getRowsCount(ctx.getSchemaName(), "clients"));
+
+                    indexes = indexMaintenance.getIndexesWithNullValues(PgContext.of("; select pg_sleep(100000000);"));
+                    assertNotNull(indexes);
+                    assertThat(indexes, hasSize(0));
+
+                    indexes = indexMaintenance.getIndexesWithNullValues(ctx);
+                    assertNotNull(indexes);
+                    assertThat(indexes, hasSize(1));
                 });
     }
 }
