--- src/test/java/io/github/mfvanek/pg/index/maintenance/IndexMaintenanceImplTestPg10.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.index.maintenance;
-
-import com.opentable.db.postgres.junit5.EmbeddedPostgresExtension;
-import com.opentable.db.postgres.junit5.PreparedDbExtension;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
-
-class IndexMaintenanceImplTestPg10 extends IndexMaintenanceImplTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension.preparedDatabase(ds -> {});
-
-    IndexMaintenanceImplTestPg10() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-
-    @Test
-    void pgVersion() {
-        assertThat(getPgVersion(), containsString("PostgreSQL 10.6"));
-    }
-}
