--- src/test/java/io/github/mfvanek/pg/index/maintenance/IndexMaintenanceImplTestPg11.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.index.maintenance;
-
-import io.zonky.test.db.postgres.junit5.EmbeddedPostgresExtension;
-import io.zonky.test.db.postgres.junit5.PreparedDbExtension;
-import org.hamcrest.MatcherAssert;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import static org.hamcrest.Matchers.containsString;
-
-class IndexMaintenanceImplTestPg11 extends IndexMaintenanceImplTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension.preparedDatabase(ds -> {});
-
-    IndexMaintenanceImplTestPg11() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-
-    @Test
-    void pgVersion() {
-        MatcherAssert.assertThat(getPgVersion(), containsString("PostgreSQL 11.5"));
-    }
-}
