--- src/test/java/io/github/mfvanek/pg/index/maintenance/StatisticsMaintenanceImplTestPg10.java
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.index.maintenance;
-
-import com.opentable.db.postgres.junit5.EmbeddedPostgresExtension;
-import com.opentable.db.postgres.junit5.PreparedDbExtension;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-class StatisticsMaintenanceImplTestPg10 extends StatisticsMaintenanceImplTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension.preparedDatabase(ds -> {});
-
-    StatisticsMaintenanceImplTestPg10() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-}
