--- src/test/java/io/github/mfvanek/pg/settings/ConfigurationMaintenanceImplTestPg11.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2019-2020. Ivan Vakhrushev and others.
- * https://github.com/mfvanek/pg-index-health
- *
- * This file is a part of "pg-index-health" - a Java library for analyzing and maintaining indexes health in PostgreSQL databases.
- */
-
-package io.github.mfvanek.pg.settings;
-
-import io.zonky.test.db.postgres.junit5.EmbeddedPostgresExtension;
-import io.zonky.test.db.postgres.junit5.PreparedDbExtension;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-class ConfigurationMaintenanceImplTestPg11 extends ConfigurationMaintenanceImplTestBase {
-
-    @RegisterExtension
-    static final PreparedDbExtension embeddedPostgres =
-            EmbeddedPostgresExtension
-                    .preparedDatabase(ds -> {})
-                    .customize(builder -> builder.setServerConfig(ImportantParam.LOCK_TIMEOUT.getName(), "1000"));
-
-    ConfigurationMaintenanceImplTestPg11() {
-        super(embeddedPostgres.getTestDatabase());
-    }
-}
